Improvements to the post-start hook.
* Fetched service states while holding the playground lock.
* Used a tuple for all_services (cached_properties should be immutable)
* Moved get_service_state -> Service
